Move MostPopularTracks into a folder.
